refactor: add file-size governance ratchet + split gearComparisonService
Tackles CQ-03 (god files) from the code-quality audit on two fronts.

Systemic guard: scripts/governance/check-file-size-budget.mjs (wired into
npm run governance as governance:filesize) freezes the current god-file debt —
CI fails if any new source file crosses 800 lines or a baselined file grows
past its recorded ceiling. Mirrors the existing source-boundary ratchet pattern;
baseline in scripts/governance/file-size-baseline.json, regenerate with
'npm run governance:filesize -- --write-baseline'. Verified it passes at the
current baseline and fails on simulated growth.

First paydown: src/utils/gearComparisonService.ts (1,068 lines, zero tests) is
decomposed into a focused src/utils/gear-comparison/ module set (types,
availableGear, compareArtistRequirements, calculateEquipmentNeeds,
getMismatchSummary; largest module 568 lines) behind a backward-compatible
barrel, so all six existing consumers import unchanged. Added the module's first
characterization tests (7 cases) covering console availability, band-provided
gear, stage-1 vs other-stage inventory, equipment-needs aggregation, and the
mismatch summary. God-file count over 800 lines: 46 -> 45.

The split is a pure code move (no logic changes), validated by typecheck, eslint
(0 errors), the new tests, and both governance ratchets.
