+++ package.json
@@ -21,8 +21,9 @@
     "test:e2e": "playwright test --project=chromium",
     "perf:baseline": "node scripts/performance/collect-phase4-baseline.mjs",
     "budget:bundle": "node scripts/performance/check-bundle-budget.mjs",
-    "governance": "npm run governance:source && npm run governance:functions && npm run governance:exposure && npm run governance:sql-grants && npm run governance:actions && npm run ci:db:migrations && npm run audit:deps",
+    "governance": "npm run governance:source && npm run governance:filesize && npm run governance:functions && npm run governance:exposure && npm run governance:sql-grants && npm run governance:actions && npm run ci:db:migrations && npm run audit:deps",
     "governance:source": "node scripts/governance/check-source-boundaries.mjs",
+    "governance:filesize": "node scripts/governance/check-file-size-budget.mjs",
     "governance:functions": "node scripts/governance/check-edge-functions.mjs",
     "governance:exposure": "node scripts/governance/check-edge-function-exposure.mjs",
     "governance:sql-grants": "node scripts/governance/check-security-definer-grants.mjs",
