+++ scripts/governance/file-size-baseline.json
@@ -1,0 +1,53 @@
+{
+  "version": 1,
+  "threshold": 800,
+  "generatedAt": "2026-06-28T08:57:46.188Z",
+  "note": "Line-count ceilings for files over the threshold. CI fails when a file exceeds its ceiling or a new file crosses the threshold. Regenerate with: npm run governance:filesize -- --write-baseline",
+  "files": {
+    "src/features/tour-ops/TourOpsManagementHub.tsx": 2116,
+    "src/features/tour-ops/tourSchedulingService.ts": 1885,
+    "src/features/technical-tools/power/consumos/useConsumosTool.ts": 1820,
+    "src/components/tours/TourDefaultsManager.tsx": 1722,
+    "src/components/jobs/cards/JobCardNew.tsx": 1438,
+    "src/components/tours/TourDateManagementDialog.tsx": 1342,
+    "src/components/matrix/StaffingCampaignPanel.tsx": 1297,
+    "src/components/soundvision/SoundVisionInteractiveMap.tsx": 1293,
+    "src/routes/app-route-manifest.tsx": 1280,
+    "src/components/festival/ArtistTable.tsx": 1253,
+    "src/utils/rates-pdf-export.ts": 1244,
+    "src/components/matrix/AssignJobDialog.tsx": 1238,
+    "src/pages/PayoutsDueFortnights.tsx": 1225,
+    "src/components/festival/pdf/PrintOptionsDialog.tsx": 1187,
+    "src/pages/PesosTool.tsx": 1150,
+    "src/components/timesheet/TimesheetView.tsx": 1137,
+    "src/components/festival/ArtistRequirementsForm.tsx": 1134,
+    "src/pages/TourManagement.tsx": 1118,
+    "src/features/technical-tools/power/consumos/ConsumosToolPage.tsx": 1065,
+    "src/components/hoja-de-ruta/ModernHojaDeRuta.tsx": 1056,
+    "src/components/technician/details-modal/DetailsModalTabs.tsx": 991,
+    "src/components/technician/TimesheetView.tsx": 980,
+    "src/features/wallboard/WallboardDisplay.tsx": 930,
+    "src/utils/artistPdfExport.ts": 928,
+    "src/components/matrix/TechnicianRow.tsx": 926,
+    "src/utils/rfIemTablePdfExport.ts": 924,
+    "src/components/department/EnhancedJobDetailsModal.tsx": 911,
+    "src/lib/unified-subscription-manager.ts": 911,
+    "src/pages/GlobalTasks.tsx": 898,
+    "src/utils/pdf/festivalPdfGenerator.ts": 892,
+    "src/utils/artistTablePdfExport.ts": 890,
+    "src/components/jobs/JobExpensesPanel.tsx": 882,
+    "src/components/disponibilidad/StockCreationManager.tsx": 875,
+    "src/components/dashboard/CalendarSection.tsx": 865,
+    "src/pages/festival-management/FestivalManagementView.tsx": 865,
+    "src/hooks/useOptimizedAuth.tsx": 846,
+    "src/components/feedback/AdminPanel.tsx": 844,
+    "src/hooks/useGlobalTaskMutations.ts": 843,
+    "src/components/festival/PushToFlexPullsheetDialog.tsx": 835,
+    "src/components/sound/AmplifierTool.tsx": 832,
+    "src/components/jobs/cards/job-card-new/JobCardNewView.tsx": 822,
+    "src/components/technician/TechnicianRfTableModal.tsx": 819,
+    "src/pages/FestivalArtistManagement.tsx": 815,
+    "src/components/jobs/JobAssignmentDialog.tsx": 811,
+    "src/services/flexWorkOrders.ts": 809
+  }
+}
